==> analysis/32_20230504_DM_DNAFISH_ctrl/03_seg-multiple.py <==
import skimage.io as skio
import napari
import shared.segmentation as seg
import shared.objects as obj
import tifffile as tif
import matplotlib.pyplot as plt
import shared.display as dis
from skimage.measure import label, regionprops
from skimage.filters import threshold_otsu, threshold_local, sobel
from skimage.morphology import extrema, binary_dilation, binary_erosion, disk, dilation
from skimage import segmentation
import math
import shared.image as ima
import numpy as np
import os

# INPUT PARAMETERS
# file info
master_folder = "/Users/xwyan/Dropbox/LAB/ChangLab/Projects/Data/20230504_analysis_DM_DNAFISH_ctrl/"
data_dir = "%sdata/" % master_folder
output_dir = "%sfigures/" % master_folder

exp = '20230409_DM_plate_DNAFISH_control'
row = 'C'
# column = [4, 5, 6]
column = [9, 10, 11]
sample_lst = ['%s%s' % (row, x) for x in column]

for sample in sample_lst:
    file_name = '%s_%s_RAW' % (exp, sample)
    img_hoechst_stack = skio.imread("%s/%s_ch01.tif" % (data_dir, file_name), plugin="tifffile")
    img_DNAFISH_stack = skio.imread("%s/%s_ch00.tif" % (data_dir, file_name), plugin="tifffile")

    # set parameters
    pixel_size = 58.7  # nm (sp8 confocal 3144x3144)
    cell_avg_size = 10  # um (Colo)
    nuclear_size_range = [0.6, 1.5]  # used to filter nucleus
    convex_conversion_threshold = 0.85
    local_size = 200

    # segmentation
    local_factor_nuclear = 99  # ~99, needs to be odd number, rmax if (rmax % 2 == 1) else rmax+1
    min_size_nuclear = (nuclear_size_range[0] * cell_avg_size * 1000 / (pixel_size * 2)) ** 2 * math.pi
    max_size_nuclear = (nuclear_size_range[1] * cell_avg_size * 1000 / (pixel_size * 2)) ** 2 * math.pi

    for fov in range(img_hoechst_stack.shape[0]):
        """viewer = napari.Viewer()
        viewer.add_image(img_hoechst[i, :, :], blending='additive', colormap='blue', contrast_limits=[0, img_hoechst.max()])
        viewer.add_image(img_DNAFISH[i, :, :], blending='additive', colormap='green', contrast_limits=[0, img_DNAFISH.max()])
        napari.run()"""

        print("analyzing %s, fov: %s" % (sample, fov))
        img_nuclear = img_hoechst_stack[fov, :, :]
        img_DNAFISH = img_DNAFISH_stack[fov, :, :]

        # nuclear seg
        img_nuclear_seg = seg.nuclear_seg1(img_nuclear, local_factor=local_factor_nuclear, min_size=min_size_nuclear,
                                           max_size=max_size_nuclear)

        img_nuclear_seg_convex = obj.label_resort(seg.obj_to_convex_filter(img_nuclear_seg,
                                                                           threshold=convex_conversion_threshold))

        if not os.path.exists("%s/seg_tif/%s/" % (output_dir, sample)):
            os.makedirs("%s/seg_tif/%s/" % (output_dir, sample))
        tif.imwrite("%s/seg_tif/%s/%s_%s_seg.tif" % (output_dir, sample, sample, fov), img_nuclear_seg_convex)

        # ecDNA segmentation
        img_DNAFISH_seg1 = np.zeros_like(img_DNAFISH)
        img_DNAFISH_seg1[img_DNAFISH > 12000] = 1
        img_DNAFISH_seg1 = obj.remove_small(img_DNAFISH_seg1, 20)

        tif.imwrite("%s/seg_tif/%s/%s_%s_ecseg1.tif" % (output_dir, sample, sample, fov), img_DNAFISH_seg1)

        if not os.path.exists("%s/color_img/%s/" % (output_dir, sample)):
            os.makedirs("%s/color_img/%s/" % (output_dir, sample))

        viewer = napari.Viewer()
        viewer.add_image(img_nuclear, blending='additive', colormap='blue', contrast_limits=[0, img_nuclear.max()])
        viewer.add_image(img_DNAFISH, blending='additive', colormap='green', contrast_limits=[0, img_DNAFISH.max()])
        plt.imsave('%s/color_img/%s/%s_%s_img.tiff' % (output_dir, sample, sample, fov), dis.blending(viewer))
        viewer.close()

        viewer = napari.Viewer()
        viewer.add_image(img_nuclear_seg_convex, blending='additive', colormap='blue', contrast_limits=[0, 1])
        viewer.add_image(img_DNAFISH_seg1, blending='additive', colormap='green', contrast_limits=[0, 1])
        plt.imsave('%s/color_img/%s/%s_%s_seg.tiff' % (output_dir, sample, sample, fov), dis.blending(viewer))
        viewer.close()

print("DONE!")
